MPSPHOENIX-30: implement metrics collection for binding requests per service
Since only the successful and still current binding requests can be retrieved from the database, a total binding requests and sucessful binding requests
counter have been implemented, which are ephemeral, though. This means that once the service broker is restarted, the counters are reset to zero.
This is why a successful binding request counter has been implemented rather than retrieving the successful binding from the database since the number of
successful bindings quickly gets higher than the number of binding requests of the lifetime of a service broker"