chore: Create dependency management in gradle
In the past I had suffered from classpath bugs hard to detect and fix
because the dependencies were just copy pasted from module to module. To
avoid this, and due to the lack of a standard dependency management in
gradle (as for example, in maven) I centralized the definiton and
declaration of all dependencies in gradle/dependencies.gradle .

Dependencies should be only declared in that file. In the dependency
section of the modules we will only refer to dependencies defined on
that file.

With this approach:
- It will be easy to spot mismatching versions of dependencies
- It is easier to document those dependencies
- We avoid declaring two different versions of the same dependency in
different modules
- For frameworks, like spring-boot, is easier to maintain the
consistency of the version between packages