refactor: Move all reusable classes to new module
We are trying to be able to reuse most of the classes used in the broker
Spring boot application. For that, we are dividing the module in two:

- :broker:core which can be used in other Spring boot environments
- :broker which is the specific Spring boot app developed in OSB

This way you can include in your Spring boot application a dependency
with broker-core without having problems in your classpath with all the
Spring configurations and spring boot application of this OSB broker.