chore: Add nexus-publish gradle plugin
This plugin creates each time is run a nexus staging repository: this
way if there is any error or problem, although the staging repository
remains in a open locked state, you can continue launching the task in
a new staging repository. The failed staging repository is supposed to
be cleaned by a automated task in the nexus server.

Use 'gradle publishToNexus closeRepository' for creating a staging
repository for testing and checking the upload of files.

Use 'gradle publishToNexus closeAndReleaseRepository' for creating a staging
repository which will be closed and RELEASED to maven central after
the build finished.