fix: gradle publish with bintray and mavencentral (#235)
fix: gradle publish with bintray and mavencentral

we faced this problem: the old plugin creates each time is run a nexus 
staging repository: this way if there is any error or problem, although
the staging repository remains in a open locked state, you can continue 
launching the task in a new staging repository. The failed staging repository
is supposed to be cleaned by a automated task in the nexus server.

The nebula plugin had some problems with the nexus-staging and
nexus-publishing plugins. Besides that, we were a little bit tired of
the staging repositories problems at oss.sonatype, and of using their
interface.

So we decided to move to bintray:
- Create a bintray user that should be part of swisscom organization
- You now need a 'bintrayUser' and 'bintrayApiKey' in your project
(added in gradle user home, or as enviroment variables)
- For publishing in bintray just do the common 'gradle publish'

It looks that:
- plugin definition in multimodule projects changed a little bit in
gradle
- nebula plugins are designed for publishing with autosigning in bintray
- we were not obtaining any benefit from complicating the build with
certain plugins.

With this build.gradle we are publishing and signing all modules and all
the different artifacts: javadoc, source and test.

* fix: Change circle-ci from oss.sonatype to bintray

We stopped using oss.sonatype directly and now we are publishing at
https://dl.bintray.com/swisscom/open-service-broker

I have added BINTRAY_USER and BINTRAY_API_KEY variables to circle-ci
(https://circleci.com/gh/swisscom/open-service-broker/edit#env-vars) and
I'm reading them from gradle (for not polluting the command line with
property passing). Now, circle-ci should publish to bintray and bintray
will manage to publish in oss.sonatype.org and then on mavencentral.