fix: Rename 'spring.credhub.enable` to osb.c.e
If you are using `spring.credhub` property path, Spring tries to
configure credhub beans. And if enable=false, first it configures beans
that are NOT going to be used and that probably are not correctly
configured.

Changing the name of the watched property in the @OnProperty makes all
the problems go away.