refactor: Select Credential Store by conf
The previous implementation was forcing to load all the credhub
configuration even when was not used. With this refactoring you can
select in the configuration which credential store to use ('default' if
the property is missing):

osb:
  credential:
    store: "credhub"

Apart of that, refactored classes to their real name, deleted useless
factory and added some documentation to public interfaces and
implementations.