Feature/credential store by configuration (#258)
* fix: Move c.s.c.s.b.b.c.enable to o.b.c.enable

Again, like in the case of osb.credhub.enable, if you use that enable
property in BoshCredHubTemplate, you are
first loading the credhub stuff and then enabling or disabling it. In
the case of disabling it does not make sense to be forced to configure
first.

* docs: Correct last version link in core

* refactor: Select Credential Store by conf

The previous implementation was forcing to load all the credhub
configuration even when was not used. With this refactoring you can
select in the configuration which credential store to use ('default' if
the property is missing):

osb:
  credential:
    store: "credhub"

Apart of that, refactored classes to their real name, deleted useless
factory and added some documentation to public interfaces and
implementations.