test: test BoshFacade with WireMock
bosh interaction is until now only tested with heavy usage of mocking.
With WireMock the interaction can be recorded and therefore tested with
actual REST responses.

Using immutables where suitable to improve intention and simplify code
reasoning.

Cleaning BoshClient API, `setConfig(request)` is replaced with
`postConfig(request)`.

By adding log statements and Asserts for parameters and some responses
in BoshRestClient, failures during authentication should be easier
detectable.

The BoshTemplateFactory was not necessary and simply called a new BoshTemplate.
This will simplify the BoshFacade by reducing the number of parameters.

We are getting rid of KubernetesTemplateManager and KubernetesTemplate,
and we are bringing their API to TemplateConfig.

The TemplateConfig is part of every Bosh service, therefore add it
directly to BoshBasedConfig.

Also, remove useless bosh factories BoshFacadeFactory and
BoshClientFactory, as they only added extra steps to create objects.

To simplify testing, provide a constructor for AsyncServiceProvider
rather than named autowiring.

Make TemplateConfig part of AbstractKubernetesServiceConfig, just like
for BoshBasedServices. With that ServiceTemplates is now part of the
service (e.g. redis) configuration.