refactor: Extract BoshFacade to its own module (#282)
* refactor: Extract BoshFacade to its own module

Inside core we want to be able to depend in BoshFacade without having
the rest of the OSB burden; but for doing that:
- We have to move a lot of classes to model module (most of them belong
there because their semantics)
- We have to create a new credhub-service module also clean of
dependencies with core
- WE had to modify the API of BoshFacade to use model classes instead of
the LastJobContext of core

* chore: Add a flyway module

The model module will need the flyway scripts to be applied in the
database before being used in a broker project. So adding this module we
could depend on it , and having flyway in the classpath the scripts will
be applied.

Move "dtos" used in service definition to models module, because they
only define datastructure and can be used outside the core module.

* feat: Add method in BoshFacade for checking task state

There are many semantic errors in isBoshTaskSuccessful, so it is better
if we expose the task state.