feat: Add BoshWebClient, substitute of BoshRestClient
The new REST client of BOSH use the more modern Spring WebClient.

NOTE: I tried to mimic the same behaviour as the old BoshRestClient:
- Not any kind of oAuth token expiration checking
- Going to uaa before requesting anything from BOSH

I can't say if that is good or bad at the moment, but I avoided to
change that behaviour in case it affects our current production
services.

Features:
- Provide more access to BOSH director API: I'm providing methods for
accessing the tasks associated with certain
deployment or all the deployments on that BOSH director instance; which
I suspect will be interesting for debugging purposes in the future.
- Add method to get the task with all events: It will be useful for the
future, and not only for testing; also as part
of the way of implementing business flows.

Test notes:
- Don't check timestamp in certain task states: In certain bosh task
states the timestamp is not always set, therefore
we do not test the timestamp in the task states PROCESSING and QUEUED.