feat: Introduce BoshDirectorService
I have to deprecate BoshFacade: it returns in some of its public API
collections of ServiceDetail, which is an entity in our own OSB model
that is not tied to BOSH concepts, and that could be problematic to
use as payload of messages in a messaging system.

So using the concepts created in BoshWebClient I have created
BoshDirectorService: it offers with less methods same functionality as
BoshFacade, and because its return types, is more flexible than old
BoshFacade.

The simple implementation is using BoshWebClient. The tests try to imitate
more or less the same behavior that BoshFacade was having, although API
and exceptions thrown are not the same. So for using this implementation
instead of BoshFacade you should be aware of the changes!

Changes:
- We are not using shuffleAzs functionality (Simon checked that we are
setting that property to false in all configurations we have), so I
deleted it.
- The instanceCount is now implemented slightly differently in java
version of BoshTemplate
- Define the Spring Bean BoshWebClient: I used a @Configuration class,
so BoshWebClient could be used without the need of a Spring Context
also.
- Added BoshDirectorProperties as Spring config: I have included this
bean for easy configuration of BoshWebClient; you
will only have to inject that bean in some factory method of a Spring
@Configuration and call to the boshWebClient builder method of
BoshWebClient.