feat: Use Freemarker instead of BoshTemplate
It is a much better way of dealing with templates: they are cached, we
can distribute the template in the jar and the template language allows
to do much more things than simple replacement.

I have included a simple bosh-template and a complex bosh template which
can generates most of the bosh deployments yaml that we are using in our
interactions with BOSH.

Features:
- Add environment properties to instance group: Following the same idea
 as in job properties; we use snake yaml
processor because these fields could have a really different structure
depending in which options are set.
- Pass BoshInfo to the deployment template: At the moment we are only
taking the UUID of BoshInfo, but maybe it is useful in the future.

Test Notes:
- Ignore BoshFacadeTest: We are deprecating BoshFacade, and just for
updating the mapping files
of this test you need to wait several minutes. At the moment the test
was failing but I don't mind because BoshFacade will be deleted in next
major release.
- Add fake properties to circleci: Gradle will fail if those properties
are not set. As we are using wiremock in tests we could set them to
fake.
- Fix integration tests: It looks that for some reason the
BoshDirectorProperties bean is not
being correctly initialized and was launching a NPE. I protected the
class from it, but why the configuration is not applied is not fixed.