feat: Add a cloud configure template for future
Because BOSH has a different API for posting cloud configs than the one
used for deployments, we can't follow directly the same strategy. You
need to post the content as a JSON, so you need an intermediate object
to post the config with a field with the yaml inside, called 'content'.

Just adding the template and the method that we need in a (near) future.