fix: Set eager fetch type for plans
In a multithreaded environment we won't have access to the initial
session context that loaded this entity from database. So I added a
method for eager fetching the plan with its parameters and I modififed
the ServiceInstance relationship with Plan to be eager fetched
(otherwise metric calculation will fail in a different thread).