fix: Rationalize relationships for bindings and instances
In the case of the relationship of instances with bindings, we will need
to fetch the bindings of an instance when we are in different threads,
so included the left join. I have also changed fro List to Set because
otherwise Hibernate can't fetch two collections eagerly (we should use
Set instead of List always, but I can't change List<Details> because
there are dependencies to the List interface in other classes :-()

In the case of bindings I eagerly get the to-one relationships, cascade
all the details (because they don't make sense without being associated
to a binding) and added method for eagerly fetching Details (which we
need for sure when accessing from different threads).