fix: store service detail updates only through freshly loaded si.
Due to the change of the cascade.all on the service instance, it happened
that some code overwrote itself, in doing so it saved a change and then
wrote immediately an delete. This has now been fixed in LocalInventoryServiceImpl and
ProvisioingPersistenceService by loading the service instance before doing
changes on the service details.