fix: REturn empty BoshDeployment when 4XX or 5XX
When BOSH server is returning a 4XX when trying to delete
is because the deployment does not
exist. If returns a 5XX is a BOSH server problem and we can't do
anything about. So the logical return should be an empty deployment, and
let each client of the API decide if that is a problem or not.