feat: Modularize the metrics in its own module
We want to be able to deactivate the metrics gathering if there is any
problem with the metrics servers. The easiest way in this moment was
creating a separated module.

WARNING: For modularizing we had to delete any reference to
PlanBasedMetricsService in
com.swisscom.cloud.sb.broker.servicedefinition.ServiceDefinitionProcessor
so that means that metrics gathering is BROKEN with this commit!