refactor: Introduce configuration module
To allow more freedom to consumers of the open-service-broker
the configuration module is introduced. Consumers can but do
not have to use this module, which allows the configuration of
services.

To allow the split of configuration and implementation, the
credential-service module needs to be introduced as well.
This prevents a circular-dependency between :broker:core and
:broker:core:configuration.