bug: Update Should Be Persisted After Success
In the case of synchronous services updating a service
can be persisted as it is now in the UpdatingService.
For asynchronous services however it's problematic to
update the service instance on request, as it might fail
at a later stage and the service instance entry is then
in a wrong state.